refactoring more helper methods
