+++ shared/src/persistence/awsDynamoPersistence.js
@@ -1,20 +1,6 @@
 const client = require('./dynamodbClientService');
 
-const stripInternalKeys = items => {
-  const strip = item => {
-    delete item.sk;
-    delete item.pk;
-  };
-  if (!items) {
-    return null;
-  } else if (items.length) {
-    items.forEach(strip);
-  } else {
-    strip(items);
-  }
-  return items;
-};
-exports.stripInternalKeys = stripInternalKeys;
+const { stripInternalKeys } = require('./dynamo/helpers/stripInternalKeys');
 
 const getSortRecordsViaMapping = async ({
   applicationContext,
@@ -130,37 +116,6 @@ const getRecordViaMapping = async ({
 
 exports.getRecordViaMapping = getRecordViaMapping;
 
-/**
- * incrementCounter
- * @param applicationContext
- * @returns {*}
- */
-// TODO: Hard coded to update the docketNumberCounter, but should
-// be refactored to update any type of atomic counter in dynamo
-// if we can pass in table / key
-exports.incrementCounter = ({ applicationContext }) => {
-  const TABLE = `efcms-${applicationContext.environment.stage}`;
-
-  const year = new Date().getFullYear().toString();
-
-  return client.updateConsistent({
-    applicationContext,
-    TableName: TABLE,
-    Key: {
-      pk: `docketNumberCounter-${year}`,
-      sk: `docketNumberCounter-${year}`,
-    },
-    UpdateExpression: 'ADD #a :x',
-    ExpressionAttributeNames: {
-      '#a': 'id',
-    },
-    ExpressionAttributeValues: {
-      ':x': 1,
-    },
-    ReturnValues: 'UPDATED_NEW',
-  });
-};
-
 const createRespondentCaseMapping = async ({
   applicationContext,
   caseId,
@@ -177,72 +132,3 @@ const createRespondentCaseMapping = async ({
 };
 
 exports.createRespondentCaseMapping = createRespondentCaseMapping;
-
-exports.deleteMappingRecord = async ({
-  applicationContext,
-  pkId,
-  skId,
-  type,
-}) => {
-  await client.delete({
-    applicationContext,
-    tableName: `efcms-${applicationContext.environment.stage}`,
-    key: {
-      pk: `${pkId}|${type}`,
-      sk: skId,
-    },
-  });
-};
-
-exports.createMappingRecord = async ({
-  applicationContext,
-  pkId,
-  skId,
-  type,
-}) => {
-  return applicationContext.getDynamoClient().put({
-    applicationContext,
-    TableName: `efcms-${applicationContext.environment.stage}`,
-    Item: {
-      pk: `${pkId}|${type}`,
-      sk: skId,
-    },
-  });
-};
-
-exports.createSortMappingRecord = async ({
-  applicationContext,
-  pkId,
-  skId,
-  item,
-  type,
-}) => {
-  return client.put({
-    applicationContext,
-    TableName: `efcms-${applicationContext.environment.stage}`,
-    Item: {
-      pk: `${pkId}|${type}`,
-      sk: skId,
-      ...item,
-    },
-  });
-};
-
-const stripWorkItems = (casesToModify, isAuthorizedForWorkItems) => {
-  if (isAuthorizedForWorkItems) return casesToModify;
-  if (!casesToModify) return casesToModify;
-
-  const strip = caseToModify => {
-    delete caseToModify.workItems;
-  };
-
-  if (casesToModify.length) {
-    casesToModify.forEach(strip);
-    return casesToModify;
-  } else {
-    strip(casesToModify);
-    return casesToModify;
-  }
-};
-
-exports.stripWorkItems = stripWorkItems;
