+++ shared/src/persistence/dynamo/cases/getCasesByUser.js
@@ -1,8 +1,7 @@
-const {
-  getRecordsViaMapping,
-  stripInternalKeys,
-  stripWorkItems,
-} = require('../../awsDynamoPersistence');
+const { getRecordsViaMapping } = require('../../awsDynamoPersistence');
+const { stripWorkItems } = require('../../dynamo/helpers/stripWorkItems');
+
+const { stripInternalKeys } = require('../../dynamo/helpers/stripInternalKeys');
 
 exports.getCasesByUser = async ({ user, applicationContext }) => {
   const cases = await getRecordsViaMapping({
