+++ shared/src/persistence/dynamo/documents/syncDocuments.js
@@ -1,4 +1,6 @@
-const persistence = require('../../awsDynamoPersistence');
+const {
+  createMappingRecord,
+} = require('../../dynamo/helpers/createMappingRecord');
 
 exports.syncDocuments = async ({
   applicationContext,
@@ -10,7 +12,7 @@ exports.syncDocuments = async ({
       i => i.documentId === document.documentId,
     );
     if (!existing) {
-      await persistence.createMappingRecord({
+      await createMappingRecord({
         applicationContext,
         pkId: document.documentId,
         skId: caseToSave.caseId,
