+++ shared/src/persistence/dynamo/users/getUsersInSection.js
@@ -1,5 +1,5 @@
 const client = require('../../dynamodbClientService');
-const { stripInternalKeys } = require('../../awsDynamoPersistence');
+const { stripInternalKeys } = require('../../dynamo/helpers/stripInternalKeys');
 
 exports.getUsersInSection = async ({ applicationContext, section }) => {
   const TABLE = `efcms-${applicationContext.environment.stage}`;
